Unify tests for different Processors
