+++ python/feathub/feature_tables/tests/test_datagen_source.py
@@ -11,28 +11,19 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#      https://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
 import unittest
+from abc import ABC
 from typing import cast
 
-from feathub.common.types import Int64
+from feathub.common.types import Int32, Timestamp, Int64
 from feathub.feature_tables.sources.datagen_source import (
-    DEFAULT_BOUNDED_NUMBER_OF_ROWS,
     DataGenSource,
+    RandomField,
     SequenceField,
+    DEFAULT_BOUNDED_NUMBER_OF_ROWS,
 )
 from feathub.table.schema import Schema
+from feathub.tests.feathub_it_test_base import FeathubITTestBase
 
 
 class DataGenSourceTest(unittest.TestCase):
@@ -75,3 +66,23 @@ class DataGenSourceTest(unittest.TestCase):
         bounded_source_json.pop("number_of_rows")
 
         self.assertEqual(source_json, bounded_source_json)
+
+
+class DataGenSourceITTest(ABC, FeathubITTestBase):
+    def test_data_gen_source(self):
+        source = DataGenSource(
+            name="datagen_src",
+            rows_per_second=10,
+            field_configs={
+                "id": SequenceField(start=0, end=9),
+                "val": RandomField(minimum=0, maximum=100),
+            },
+            schema=Schema(["id", "val", "ts"], [Int32, Int32, Timestamp]),
+            timestamp_field="ts",
+            timestamp_format="%Y-%m-%d %H:%M:%S",
+        )
+
+        df = self.client.get_features(features=source).to_pandas()
+
+        self.assertEquals(10, df.shape[0])
+        self.assertTrue((df["val"] >= 0).all() and (df["val"] <= 100).all())
