+++ python/feathub/feature_views/transforms/tests/test_expression_transform.py
@@ -11,23 +11,23 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from feathub.processors.spark.spark_processor import SparkProcessor
+from abc import ABC
 
 from feathub.common.types import Float64
 from feathub.feature_views.derived_feature_view import DerivedFeatureView
 from feathub.feature_views.feature import Feature
-from feathub.processors.spark.tests.spark_test_utils import SparkProcessorTestBase
+from feathub.tests.feathub_it_test_base import FeathubITTestBase
 
 
-class SparkDerivedFeatureViewTest(SparkProcessorTestBase):
-    def test_derived_feature_view(self):
-        processor = SparkProcessor(
-            props={"processor.spark.master": "local[1]"},
-            registry=self.registry,
-        )
+class ExpressionTransformITTest(ABC, FeathubITTestBase):
+    def test_expression_transform(self):
+        self._test_expression_transform(False)
 
-        source = self._create_file_source(self.input_data.copy(), schema=self.schema)
+    def test_expression_transform_keep_source_fields(self):
+        self._test_expression_transform(True)
+
+    def _test_expression_transform(self, keep_source_fields: bool):
+        source = self.create_file_source(self.input_data.copy())
 
         f_cost_per_mile = Feature(
             name="cost_per_mile",
@@ -41,25 +41,26 @@ class SparkDerivedFeatureViewTest(SparkProcessorTestBase):
             features=[
                 f_cost_per_mile,
             ],
-            keep_source_fields=False,
+            keep_source_fields=keep_source_fields,
         )
 
-        result_df = (
-            processor.get_table(features)
-            .to_pandas()
-            .sort_values(by=["name", "time"])
-            .reset_index(drop=True)
-        )
+        result_df = self.client.get_features(features).to_pandas()
 
-        expected_result_df = self.input_data
+        expected_result_df = self.input_data.copy()
         expected_result_df["cost_per_mile"] = expected_result_df.apply(
             lambda row: row["cost"] / row["distance"] + 10, axis=1
         )
-        expected_result_df = expected_result_df.sort_values(
-            by=["name", "time"]
-        ).reset_index(drop=True)
-        expected_result_df["cost"] = expected_result_df["cost"].astype("int32")
-        expected_result_df["distance"] = expected_result_df["distance"].astype("int32")
+
+        if keep_source_fields:
+            result_df = result_df.sort_values(by=["name", "time"])
+            expected_result_df = expected_result_df.sort_values(by=["name", "time"])
+        else:
+            result_df = result_df.sort_values(by=["time"])
+            expected_result_df.drop(["name", "cost", "distance"], axis=1, inplace=True)
+            expected_result_df = expected_result_df.sort_values(by=["time"])
+
+        result_df = result_df.reset_index(drop=True)
+        expected_result_df = expected_result_df.reset_index(drop=True)
 
         self.assertIsNone(source.keys)
         self.assertIsNone(features.keys)
