+++ python/feathub/feature_views/transforms/tests/test_over_window_transform.py
@@ -11,6 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+from abc import ABC
 from datetime import timedelta
 from math import sqrt
 
@@ -21,14 +22,101 @@ from feathub.feature_views.derived_feature_view import DerivedFeatureView
 from feathub.feature_views.feature import Feature
 from feathub.feature_views.transforms.over_window_transform import OverWindowTransform
 from feathub.feature_views.transforms.python_udf_transform import PythonUdfTransform
-from feathub.processors.flink.flink_table import flink_table_to_pandas
-from feathub.processors.flink.table_builder.tests.table_builder_test_utils import (
-    FlinkTableBuilderTestBase,
-)
 from feathub.table.schema import Schema
+from feathub.tests.feathub_it_test_base import FeathubITTestBase
 
 
-class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
+class OverWindowTransformITTest(ABC, FeathubITTestBase):
+    def test_over_window_transform(self):
+        df = self.input_data.copy()
+        source = self.create_file_source(df)
+
+        f_cost_per_mile = Feature(
+            name="cost_per_mile",
+            dtype=Float64,
+            transform="CAST(cost AS DOUBLE) / CAST(distance AS DOUBLE) + 10",
+        )
+
+        f_total_cost = Feature(
+            name="total_cost",
+            dtype=Int64,
+            transform=OverWindowTransform(
+                expr="cost",
+                agg_func="SUM",
+                group_by_keys=["name"],
+                window_size=timedelta(days=2),
+            ),
+        )
+        f_avg_cost = Feature(
+            name="avg_cost",
+            dtype=Float64,
+            transform=OverWindowTransform(
+                expr="cost",
+                agg_func="AVG",
+                group_by_keys=["name"],
+                window_size=timedelta(days=2),
+            ),
+        )
+        f_max_cost = Feature(
+            name="max_cost",
+            dtype=Int64,
+            transform=OverWindowTransform(
+                expr="cost",
+                agg_func="MAX",
+                group_by_keys=["name"],
+                window_size=timedelta(days=2),
+            ),
+        )
+        f_min_cost = Feature(
+            name="min_cost",
+            dtype=Int64,
+            transform=OverWindowTransform(
+                expr="cost",
+                agg_func="MIN",
+                group_by_keys=["name"],
+                window_size=timedelta(days=2),
+            ),
+        )
+
+        features = DerivedFeatureView(
+            name="feature_view",
+            source=source,
+            features=[
+                f_cost_per_mile,
+                f_total_cost,
+                f_avg_cost,
+                f_max_cost,
+                f_min_cost,
+            ],
+            keep_source_fields=False,
+        )
+
+        result_df = (
+            self.client.get_features(features=features)
+            .to_pandas()
+            .sort_values(by=["name", "time"])
+            .reset_index(drop=True)
+        )
+
+        expected_result_df = df
+        expected_result_df["cost_per_mile"] = expected_result_df.apply(
+            lambda row: row["cost"] / row["distance"] + 10, axis=1
+        )
+        expected_result_df["total_cost"] = pd.Series([100, 400, 400, 600, 500, 900])
+        expected_result_df["avg_cost"] = pd.Series(
+            [100.0, 400.0, 200.0, 300.0, 500.0, 450.0]
+        )
+        expected_result_df["max_cost"] = pd.Series([100, 400, 300, 400, 500, 600])
+        expected_result_df["min_cost"] = pd.Series([100, 400, 100, 200, 500, 300])
+        expected_result_df.drop(["cost", "distance"], axis=1, inplace=True)
+        expected_result_df = expected_result_df.sort_values(
+            by=["name", "time"]
+        ).reset_index(drop=True)
+
+        self.assertIsNone(source.keys)
+        self.assertListEqual(["name"], features.keys)
+        self.assertTrue(expected_result_df.equals(result_df))
+
     def test_over_window_transform_with_unsupported_agg_func(self):
         with self.assertRaises(ValueError):
             Feature(
@@ -41,7 +129,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
 
     def test_over_window_transform_without_key(self):
         df = self.input_data.copy()
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         f_total_cost = Feature(
             name="total_cost",
@@ -67,7 +155,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         ).reset_index(drop=True)
 
         result_df = (
-            self.flink_table_builder.build(features=features)
+            self.client.get_features(features=features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -76,7 +164,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
 
     def test_over_window_transform_without_window_size_and_limit(self):
         df = self.input_data.copy()
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         f_total_cost = Feature(
             name="total_cost",
@@ -100,7 +188,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         )
 
         result_df = (
-            self.flink_table_builder.build(features)
+            self.client.get_features(features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -109,7 +197,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
 
     def test_over_window_transform_with_limit(self):
         df = self.input_data.copy()
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         f_total_cost = Feature(
             name="total_cost",
@@ -133,27 +221,19 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         )
 
         result_df = (
-            self.flink_table_builder.build(features)
+            self.client.get_features(features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
         )
         self.assertTrue(expected_result_df.equals(result_df))
 
-    def test_over_window_transform_with_window_size(self):
-        df = pd.DataFrame(
-            [
-                ["Alex", 100, 100, "2022-01-01 08:00:00.001"],
-                ["Emma", 400, 250, "2022-01-01 08:00:00.002"],
-                ["Alex", 300, 200, "2022-01-01 08:00:00.003"],
-                ["Emma", 200, 250, "2022-01-01 08:00:00.004"],
-                ["Jack", 500, 500, "2022-01-01 08:00:00.005"],
-                ["Alex", 600, 800, "2022-01-01 08:00:00.006"],
-            ],
-            columns=["name", "cost", "distance", "time"],
-        )
+    def test_over_window_transform_with_millis_window_size(self):
+        df, schema = self.create_input_data_and_schema_with_millis_time_span()
 
-        source = self._create_file_source(df, timestamp_format="%Y-%m-%d %H:%M:%S.%f")
+        source = self.create_file_source(
+            df, timestamp_format="%Y-%m-%d %H:%M:%S.%f", schema=schema
+        )
 
         features = DerivedFeatureView(
             name="feature_view",
@@ -180,7 +260,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         ).reset_index(drop=True)
 
         result_df = (
-            self.flink_table_builder.build(features=features)
+            self.client.get_features(features=features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -200,7 +280,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
             columns=["name", "cost", "distance", "time"],
         )
 
-        source = self._create_file_source(
+        source = self.create_file_source(
             df,
             timestamp_format="epoch_millis",
             schema=Schema.new_builder()
@@ -236,7 +316,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         ).reset_index(drop=True)
 
         result_df = (
-            self.flink_table_builder.build(features=features)
+            self.client.get_features(features=features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -244,19 +324,11 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         self.assertTrue(expected_result_df.equals(result_df))
 
     def test_expression_transform_on_over_window_transform(self):
-        df = pd.DataFrame(
-            [
-                ["Alex", 100, 100, "2022-01-01 08:00:00.001"],
-                ["Emma", 400, 250, "2022-01-01 08:00:00.002"],
-                ["Alex", 300, 200, "2022-01-01 08:00:00.003"],
-                ["Emma", 200, 250, "2022-01-01 08:00:00.004"],
-                ["Jack", 500, 500, "2022-01-01 08:00:00.005"],
-                ["Alex", 600, 800, "2022-01-01 08:00:00.006"],
-            ],
-            columns=["name", "cost", "distance", "time"],
-        )
+        df, schema = self.create_input_data_and_schema_with_millis_time_span()
 
-        source = self._create_file_source(df, timestamp_format="%Y-%m-%d %H:%M:%S.%f")
+        source = self.create_file_source(
+            df, timestamp_format="%Y-%m-%d %H:%M:%S.%f", schema=schema
+        )
 
         features = DerivedFeatureView(
             name="feature_view",
@@ -287,7 +359,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         ).reset_index(drop=True)
 
         result_df = (
-            self.flink_table_builder.build(features=features)
+            self.client.get_features(features=features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -295,19 +367,11 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         self.assertTrue(expected_result_df.equals(result_df))
 
     def test_python_udf_transform_on_over_window_transform(self):
-        df = pd.DataFrame(
-            [
-                ["Alex", 100, 100, "2022-01-01 08:00:00.001"],
-                ["Emma", 400, 250, "2022-01-01 08:00:00.002"],
-                ["Alex", 300, 200, "2022-01-01 08:00:00.003"],
-                ["Emma", 200, 250, "2022-01-01 08:00:00.004"],
-                ["Jack", 500, 500, "2022-01-01 08:00:00.005"],
-                ["Alex", 600, 800, "2022-01-01 08:00:00.006"],
-            ],
-            columns=["name", "cost", "distance", "time"],
-        )
+        df, schema = self.create_input_data_and_schema_with_millis_time_span()
 
-        source = self._create_file_source(df, timestamp_format="%Y-%m-%d %H:%M:%S.%f")
+        source = self.create_file_source(
+            df, schema=schema, timestamp_format="%Y-%m-%d %H:%M:%S.%f"
+        )
 
         features = DerivedFeatureView(
             name="feature_view",
@@ -350,11 +414,63 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         ).reset_index(drop=True)
 
         result_df = (
-            self.flink_table_builder.build(features=features)
+            self.client.get_features(features=features)
             .to_pandas()
             .sort_values(by=["lower_name", "time"])
             .reset_index(drop=True)
         )
+        self.assertTrue(expected_result_df.equals(result_df))
+
+    def test_expression_transform_and_over_window_transform(self):
+        df_1 = self.input_data.copy()
+        source = self.create_file_source(df_1)
+
+        feature_view_1 = DerivedFeatureView(
+            name="feature_view_1",
+            source=source,
+            features=[
+                Feature(
+                    "avg_cost",
+                    dtype=Float64,
+                    transform=OverWindowTransform(
+                        expr="cost",
+                        agg_func="AVG",
+                        group_by_keys=["name"],
+                        window_size=timedelta(days=2),
+                    ),
+                ),
+                Feature("10_times_cost", dtype=Int64, transform="10 * cost"),
+            ],
+            keep_source_fields=True,
+        )
+
+        feature_view_2 = DerivedFeatureView(
+            name="feature_view_2",
+            source=feature_view_1,
+            features=["avg_cost", "10_times_cost"],
+        )
+
+        self.client.build_features([feature_view_1, feature_view_2])
+
+        expected_result_df = df_1
+        expected_result_df["avg_cost"] = pd.Series(
+            [100.0, 400.0, 200.0, 300.0, 500.0, 450.0]
+        )
+        expected_result_df["10_times_cost"] = pd.Series(
+            [1000, 4000, 3000, 2000, 5000, 6000]
+        )
+        expected_result_df.drop(["cost", "distance"], axis=1, inplace=True)
+        expected_result_df = expected_result_df.sort_values(
+            by=["name", "time"]
+        ).reset_index(drop=True)
+
+        result_df = (
+            self.client.get_features("feature_view_2")
+            .to_pandas()
+            .sort_values(by=["name", "time"])
+            .reset_index(drop=True)
+        )
+
         self.assertTrue(expected_result_df.equals(result_df))
 
     def test_over_window_transform_with_window_size_and_limit(self):
@@ -374,7 +490,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         )
 
         schema = Schema(["name", "cost", "time"], [String, Float64, String])
-        source = self._create_file_source(df, schema=schema)
+        source = self.create_file_source(df, schema=schema)
 
         expected_df = df.copy()
         expected_df["last_2_last_2_minute_total_cost"] = pd.Series(
@@ -445,7 +561,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
             ],
         )
 
-        table = self.flink_table_builder.build(features=features)
+        table = self.client.get_features(features=features)
         result_df = (
             table.to_pandas().sort_values(by=["name", "time"]).reset_index(drop=True)
         )
@@ -454,7 +570,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
 
     def test_over_window_transform_first_last_value(self):
         df = self.input_data.copy()
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         feature_view = DerivedFeatureView(
             name="feature_view",
@@ -512,7 +628,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         )
 
         result_df = (
-            self.flink_table_builder.build(feature_view)
+            self.client.get_features(feature_view)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -521,7 +637,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
 
     def test_over_window_transform_row_num(self):
         df = self.input_data.copy()
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         feature_view = DerivedFeatureView(
             name="feature_view",
@@ -549,7 +665,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         )
 
         result_df = (
-            self.flink_table_builder.build(feature_view)
+            self.client.get_features(feature_view)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
@@ -569,7 +685,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
             ],
             columns=["name", "cost", "distance", "time"],
         )
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         feature_view = DerivedFeatureView(
             name="feature_view",
@@ -627,9 +743,9 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
             drop=True
         )
 
-        table = self.flink_table_builder.build(feature_view)
         result_df = (
-            flink_table_to_pandas(table)
+            self.client.get_features(feature_view)
+            .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
         )
@@ -655,7 +771,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         schema = Schema(
             ["name", "action", "cost", "time"], [String, String, Float64, String]
         )
-        source = self._create_file_source(df, schema=schema, keys=["name"])
+        source = self.create_file_source(df, schema=schema, keys=["name"])
 
         expected_df = df.copy()
         expected_df["last_2_pay_last_2_minute_total_cost"] = pd.Series(
@@ -685,7 +801,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
             ],
         )
 
-        table = self.flink_table_builder.build(features=features)
+        table = self.client.get_features(features=features)
         result_df = (
             table.to_pandas().sort_values(by=["name", "time"]).reset_index(drop=True)
         )
@@ -694,7 +810,7 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
 
     def test_over_window_transform_with_different_criteria(self):
         df = self.input_data.copy()
-        source = self._create_file_source(df)
+        source = self.create_file_source(df)
 
         f_all_total_cost = Feature(
             name="all_total_cost",
@@ -775,9 +891,97 @@ class FlinkTableBuilderOverWindowTransformTest(FlinkTableBuilderTestBase):
         ).reset_index(drop=True)
 
         result_df = (
-            self.flink_table_builder.build(features=features)
+            self.client.get_features(features=features)
             .to_pandas()
             .sort_values(by=["name", "time"])
             .reset_index(drop=True)
         )
+        self.assertTrue(expected_result_df.equals(result_df))
+
+    def test_over_window_on_join_field(self):
+        df_1 = self.input_data.copy()
+        source = self.create_file_source(df_1)
+
+        df_2 = pd.DataFrame(
+            [
+                ["Alex", 100.0, "2022-01-01,09:01:00"],
+                ["Emma", 400.0, "2022-01-01,09:02:00"],
+                ["Alex", 200.0, "2022-01-02,09:03:00"],
+                ["Emma", 300.0, "2022-01-02,09:04:00"],
+                ["Jack", 500.0, "2022-01-03,09:05:00"],
+                ["Alex", 450.0, "2022-01-03,09:06:00"],
+            ],
+            columns=["name", "avg_cost", "time"],
+        )
+        source_2 = self.create_file_source(
+            df_2,
+            schema=Schema(["name", "avg_cost", "time"], [String, Float64, String]),
+            timestamp_format="%Y-%m-%d,%H:%M:%S",
+            keys=["name"],
+        )
+        feature_view_2 = DerivedFeatureView(
+            name="feature_view_2",
+            source=source,
+            features=[
+                Feature(
+                    name="cost",
+                    dtype=Int64,
+                    transform="cost",
+                ),
+                "distance",
+                f"{source_2.name}.avg_cost",
+                Feature(
+                    name="derived_cost",
+                    dtype=Float64,
+                    transform="avg_cost * distance",
+                ),
+                Feature(
+                    name="last_avg_cost",
+                    dtype=Int64,
+                    transform=OverWindowTransform(
+                        expr="avg_cost",
+                        agg_func="LAST_VALUE",
+                        window_size=timedelta(days=2),
+                        group_by_keys=["name"],
+                        limit=2,
+                    ),
+                ),
+                Feature(
+                    name="double_last_avg_cost",
+                    dtype=Float64,
+                    transform="last_avg_cost * 2",
+                ),
+            ],
+            keep_source_fields=False,
+        )
+
+        [_, built_feature_view_2] = self.client.build_features(
+            [source_2, feature_view_2]
+        )
+
+        expected_result_df = df_1
+        expected_result_df["avg_cost"] = pd.Series(
+            [None, None, 100.0, 400.0, None, 200.0]
+        )
+        expected_result_df["derived_cost"] = pd.Series(
+            [None, None, 20000.0, 100000.0, None, 160000.0]
+        )
+        expected_result_df["last_avg_cost"] = pd.Series(
+            [None, None, 100.0, 400.0, None, 200.0]
+        )
+        expected_result_df["double_last_avg_cost"] = pd.Series(
+            [None, None, 200.0, 800.0, None, 400.0]
+        )
+        expected_result_df = expected_result_df.sort_values(
+            by=["name", "time"]
+        ).reset_index(drop=True)
+
+        result_df = (
+            self.client.get_features(features=built_feature_view_2)
+            .to_pandas()
+            .sort_values(by=["name", "time"])
+            .reset_index(drop=True)
+        )
+
+        self.assertListEqual(["name"], built_feature_view_2.keys)
         self.assertTrue(expected_result_df.equals(result_df))
