+++ python/feathub/processors/flink/job_submitter/flink_job_submitter.py
@@ -61,7 +61,7 @@ class FlinkJobSubmitter(ABC):
         :param local_registry_tables: All the table descriptors registered in the local
                                       registry that are required to compute the given
                                       table.
-        :param allow_overwrite: If it is true, throw error if the features collide with
+        :param allow_overwrite: If it is false, throw error if the features collide with
                                 existing data in the given sink.
         :return:ProcessorJob A processor job representing the submitted Flink job.
         """
