--- python/feathub/processors/local/tests/test_source.py
@@ -1,42 +1,0 @@
-# Copyright 2022 The Feathub Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import pandas as pd
-
-from feathub.common.test_utils import LocalProcessorTestCase
-
-
-class SourceTest(LocalProcessorTestCase):
-    def setUp(self):
-        super().setUp()
-        self.input_data = pd.DataFrame(
-            [
-                ["Alex", 100, 100, "2022-01-01 08:01:00"],
-                ["Emma", 400, 250, "2022-01-01 08:02:00"],
-                ["Alex", 300, 200, "2022-01-02 08:03:00"],
-                ["Emma", 200, 250, "2022-01-02 08:04:00"],
-                ["Jack", 500, 500, "2022-01-03 08:05:00"],
-                ["Alex", 600, 800, "2022-01-03 08:06:00"],
-            ],
-            columns=["name", "cost", "distance", "time"],
-        )
-
-    def tearDown(self):
-        super().tearDown()
-
-    def test_file_source(self):
-        df = self.input_data.copy()
-        source = self._create_file_source(df)
-        result_df = self.processor.get_table(features=source).to_pandas()
-        self.assertTrue(df.equals(result_df))
